Split health UI functionality out of hand absorber component
