+++ Assets/Resources/Prefabs/HealthUIInteractable.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0def0789e1190b14ea48cf9921103751
+guid: da80ce6fd7a5eb64e8f887d4a0cfe671
 PrefabImporter:
   externalObjects: {}
   userData:
